requirements.txt
annotated-types>=0.7.0
anyio>=4.6.0
certifi>=2024.8.30
chardet>=5.2.0
charset-normalizer>=3.3.2
click>=8.1.7
distro>=1.9.0
dnspython>=2.6.1
email_validator>=2.2.0
fastapi>=0.115.0
fastapi-cli>=0.0.5
fpdf>=1.7.2
h11>=0.14.0
httpcore>=1.0.5
httptools>=0.6.1
httpx>=0.27.2
idna>=3.10
Jinja2>=3.1.4
jiter>=0.5.0
loguru>=0.7.2
markdown-it-py>=3.0.0
MarkupSafe>=2.1.5
mdurl>=0.1.2
openai>=1.50.2
orjson>=3.10.7
pillow>=10.4.0
pydantic>=2.9.2
pydantic_core>=2.23.4
Pygments>=2.18.0
python-dotenv>=1.0.1
python-multipart>=0.0.12
PyYAML>=6.0.2
reportlab>=4.2.4
requests>=2.32.3
rich>=13.8.1
shellingham>=1.5.4
sniffio>=1.3.1
starlette>=0.38.6
tqdm>=4.66.5
typer>=0.12.5
typing_extensions>=4.12.2
ujson>=5.10.0
urllib3>=2.2.3
uvicorn>=0.31.0
uvloop>=0.20.0
watchfiles>=0.24.0
websockets>=13.1
pytest
streamlit
passlib
python-jose[cryptography]